requirements.txt
nltk>=3.6.2
scikit-learn>=0.24.2
gensim>=4.0.1
numpy>=1.20.3
pyenchant>=3.2.0
pandas>=1.2.5
antlr4-python3-runtime>=4.9.2
